shrink bottom panel
refactored bottom panel into its own layout and wrapped it in a view
group called TransportViewGroup. Inside this class, I declared a private
Animation class to shrink and expand this bottom panel.
Also removed the capture button from this bottom panel, so now it is
actually on top of the bottom panel and it remains static as the bottom
panel animates.